+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Adds oihana\files\getDirectory
 - Adds oihana\files\getFileExtension
 - Adds oihana\files\getHomeDirectory
+- Adds oihana\files\getOwnershipInfo
 - Adds oihana\files\getRoot
 - Adds oihana\files\getSchemeAndHierarchy
 - Adds oihana\files\getTemporaryDirectory